follow-up comments from PR #617
